pyproject.toml
[tool.poetry]
name = "coco-visualizer"
version = "0.1.0"
description = ""
authors = ["Benjamin Maccini <benjaminmaccini@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
dataclasses-json = "^0.5.2"
Pillow = "^8.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
